Committing large batch of uncommitted changes, code is in a broken state
Francesco and I are looking at the code and struggling to fix a path error, specifically in slurmPinput.py the -D parameter needs to be the absolute path and is only gettin the lowest direcotry name. So we are about to make some changes. But as we poked around we discovered that there is a whole bunch of uncommtted changes. So I am committing them first.